task_1/requirements.txt
gemini-agent-framework
python-dotenv
pdfplumber
